# [hail] memory-efficient scan