--- a/hail/src/main/scala/is/hail/HailContext.scala
+++ b/hail/src/main/scala/is/hail/HailContext.scala
@@ -271,7 +271,7 @@ object HailContext {
 
     if (!quiet)
       ProgressBarBuilder.build(sparkContext)
-     
+
     val hc = new HailContext(SparkBackend(sparkContext), logFile, tmpDir, branchingFactor, optimizerIterations)
     sparkContext.uiWebUrl.foreach(ui => info(s"SparkUI: $ui"))
 
@@ -397,7 +397,7 @@ class HailContext private(
   val sparkSession = SparkSession.builder().config(sc.getConf).getOrCreate()
   val sFS: FS = new HadoopFS(new SerializableHadoopConfiguration(sc.hadoopConfiguration))
   val bcFS: Broadcast[FS] = sc.broadcast(sFS)
-  
+
   val tmpDir = TempDir.createTempDir(tmpDirPath, sFS)
   info(s"Hail temporary directory: $tmpDir")