--- a/hail/src/main/scala/is/hail/HailContext.scala
+++ b/hail/src/main/scala/is/hail/HailContext.scala
@@ -621,7 +621,8 @@ class HailFeatureFlags {
   private[this] val flags: mutable.Map[String, String] =
     mutable.Map[String, String](
       "cpp" -> null,
-      "lower" -> null
+      "lower" -> null,
+      "max_leader_scans" -> "1000"
     )
 
   val available: java.util.ArrayList[String] =