@danking 2019-06-12
Also, cc: @akotlar I believe it were your changes that I just rebased on. Can you take a look at HadoopFS and make sure I didn't screw something up?