@danking 2019-06-13
Fixed a rebase failure, deleted some unnecessary whitespace.