@akotlar 2019-06-13
All FS implementations are not going to be Hadoop file systems. In the future apparently none of them will be.