@akotlar 2019-06-13
> t I just rebased on. Can you take a look at HadoopFS and make sure I didn't screw something up?

Let me know if you still want me to take a look whenever you reissue the pr